+++ client/doublezerod/internal/manager/http_test.go
@@ -455,7 +455,7 @@ func (m *MockNetlink) TunnelDelete(n *routing.Tunnel) error {
 	return nil
 }
 
-func (m *MockNetlink) TunnelAddrAdd(t *routing.Tunnel, ip string) error {
+func (m *MockNetlink) TunnelAddrAdd(t *routing.Tunnel, ip string, scope int) error {
 	m.tunAddrAdded = append(m.tunAddrAdded, ip)
 	return nil
 }